Create CONTRIBUTING.md
It is important, that you can give "likes" multiple times per card. (Background, if in your eyes something is really important you can push your item.
It would be nice, if you can delete your likes withour erasing the card, for example if you misclicked, or someone is explaining it in detail and you change your mind.
It is also important that the save time per board is at least 7 days by default, in case you were not able to export it on time.